[#29] Rename Android's app id
