+++ android/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="co.nimblehq.kmm.template.android">
+    package="co.nimblehq.kmm.template">
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".MyApplication"
+        android:name=".MainApplication"
         android:allowBackup="false"
         android:label="@string/app_name"
         android:supportsRtl="true"
