--- android/src/main/java/co/nimblehq/kmm/template/android/lib/TypeAlias.kt
@@ -1,3 +1,0 @@
-package co.nimblehq.kmm.template.android.lib
-
-typealias IsLoading = Boolean
