+++ android/src/main/java/co/nimblehq/kmm/template/lib/TypeAlias.kt
@@ -1,0 +1,3 @@
+package co.nimblehq.kmm.template.lib
+
+typealias IsLoading = Boolean
