+++ android/src/main/java/co/nimblehq/kmm/template/ui/screens/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.kmm.template.android.ui.screens.home
+package co.nimblehq.kmm.template.ui.screens.home
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.*
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import co.nimblehq.kmm.template.Greeting
-import co.nimblehq.kmm.template.android.ui.theme.ComposeTheme
+import co.nimblehq.kmm.template.ui.theme.ComposeTheme
 
 @Composable
 fun HomeScreen() {
